+++ lib/gruf/rspec/railtie.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# Copyright (c) 2018-present, BigCommerce Pty. Ltd. All rights reserved
+# Copyright (c) 2022-present, BigCommerce Pty. Ltd. All rights reserved
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
 # documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
@@ -15,21 +15,17 @@
 # COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
 # OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
-require_relative 'simplecov_helper'
-require 'gruf'
-require 'gruf/rspec'
-
-Dir["#{File.join(File.dirname(__FILE__), 'support')}/**/*.rb"].sort.each {|f| require f }
-
-RSpec.configure do |config|
-  config.color = true
-  config.disable_monkey_patching!
-  config.example_status_persistence_file_path = '.rspec_status'
-  config.expect_with :rspec do |c|
-    c.syntax = :expect
-  end
-  config.filter_run_excluding broken: true
-  config.mock_with :rspec do |mocks|
-    mocks.allow_message_expectations_on_nil = true
+module Gruf
+  module Rspec
+    ##
+    # Rails integration for gruf-rspec
+    #
+    class Railtie < ::Rails::Railtie
+      initializer 'gruf-rspec.initializer' do |_app|
+        config.after_initialize do
+          ::Gruf.autoloaders.load!(controllers_path: ::Gruf.controllers_path) if ::Gruf.respond_to?(:autoloaders)
+        end
+      end
+    end
   end
 end